createParty, updateParty, and deleteParty queries working
