--- DeleteTheAppQueries.go
@@ -1,60 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"net/http"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/dynamodb"
-	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbiface"
-)
-
-// Delete a Person from the database
-func deletePerson(w http.ResponseWriter, r *http.Request) {
-	facebookID := r.URL.Query().Get("facebookID")
-	queryResult := deletePersonHelper(facebookID)
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	json.NewEncoder(w).Encode(queryResult)
-}
-
-func deletePersonHelper(facebookID string) QueryResult {
-	var queryResult = QueryResult{}
-	queryResult.Succeeded = false
-	queryResult.DynamodbCalls = make([]DynamodbCall, 1)
-	type ItemGetter struct {
-		DynamoDB dynamodbiface.DynamoDBAPI
-	}
-	// Setup
-	var getter = new(ItemGetter)
-	var config = &aws.Config{Region: aws.String("us-west-2")}
-	sess, err := session.NewSession(config)
-	if err != nil {
-		queryResult.Error = "deletePersonHelper function: session creation error. " + err.Error()
-		return queryResult
-	}
-	var svc = dynamodb.New(sess)
-	getter.DynamoDB = dynamodbiface.DynamoDBAPI(svc)
-	// Finally
-	keyMap := make(map[string]*dynamodb.AttributeValue)
-	var key = dynamodb.AttributeValue{}
-	key.SetS(facebookID)
-	keyMap["facebookID"] = &key
-
-	var deleteItemInput = dynamodb.DeleteItemInput{}
-	deleteItemInput.SetTableName("Person")
-	deleteItemInput.SetKey(keyMap)
-
-	_, err2 := getter.DynamoDB.DeleteItem(&deleteItemInput)
-	var dynamodbCall = DynamodbCall{}
-	if err2 != nil {
-		dynamodbCall.Error = "deletePersonHelper function: DeleteItem error. " + err2.Error()
-		dynamodbCall.Succeeded = false
-		queryResult.DynamodbCalls[0] = dynamodbCall
-		return queryResult
-	}
-	dynamodbCall.Succeeded = true
-	queryResult.DynamodbCalls[0] = dynamodbCall
-	queryResult.Succeeded = true
-	return queryResult
-}
