+++ RateTab.go
@@ -1,0 +1,175 @@
+package main
+
+import (
+	"encoding/json"
+	"net/http"
+	"strconv"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/dynamodb"
+	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbiface"
+)
+
+// Change my rating for a party
+func rateParty(w http.ResponseWriter, r *http.Request) {
+	partyID := r.URL.Query().Get("partyID")
+	facebookID := r.URL.Query().Get("facebookID")
+	rating := r.URL.Query().Get("rating")
+	timeLastRated := r.URL.Query().Get("timeLastRated")
+	queryResult := ratePartyHelper(partyID, facebookID, rating, timeLastRated)
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	json.NewEncoder(w).Encode(queryResult)
+}
+
+// Change my rating for a bar (add my info to the attendees map if need be)
+func rateBar(w http.ResponseWriter, r *http.Request) {
+	barID := r.URL.Query().Get("barID")
+	facebookID := r.URL.Query().Get("facebookID")
+	isMale, isMaleConvErr := strconv.ParseBool(r.URL.Query().Get("isMale"))
+	name := r.URL.Query().Get("name")
+	rating := r.URL.Query().Get("rating")
+	status := "T"
+	timeLastRated := r.URL.Query().Get("timeLastRated")
+	var queryResult = QueryResult{}
+	if isMaleConvErr != nil {
+		queryResult.Error = "rateBar function: isMale parameter issue. " + isMaleConvErr.Error()
+		json.NewEncoder(w).Encode(queryResult)
+		return
+	}
+	queryResult = rateBarHelper(barID, facebookID, isMale, name, rating, status, timeLastRated)
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	json.NewEncoder(w).Encode(queryResult)
+}
+
+func ratePartyHelper(partyID string, facebookID string, rating string, timeLastRated string) QueryResult {
+	var queryResult = QueryResult{}
+	queryResult.Succeeded = false
+	queryResult.DynamodbCalls = make([]DynamodbCall, 1)
+	type ItemGetter struct {
+		DynamoDB dynamodbiface.DynamoDBAPI
+	}
+	// Setup
+	var getter = new(ItemGetter)
+	var config = &aws.Config{Region: aws.String("us-west-2")}
+	sess, err := session.NewSession(config)
+	if err != nil {
+		queryResult.Error = "ratePartyHelper function: session creation error. " + err.Error()
+		return queryResult
+	}
+	var svc = dynamodb.New(sess)
+	getter.DynamoDB = dynamodbiface.DynamoDBAPI(svc)
+	// Finally
+	expressionAttributeNames := make(map[string]*string)
+	var invitees = "invitees"
+	stringRating := "rating"
+	stringTimeLastRated := "timeLastRated"
+	expressionAttributeNames["#i"] = &invitees
+	expressionAttributeNames["#f"] = &facebookID
+	expressionAttributeNames["#r"] = &stringRating
+	expressionAttributeNames["#t"] = &stringTimeLastRated
+	expressionValuePlaceholders := make(map[string]*dynamodb.AttributeValue)
+	var ratingAttributeValue = dynamodb.AttributeValue{}
+	ratingAttributeValue.SetS(rating)
+	expressionValuePlaceholders[":rating"] = &ratingAttributeValue
+	var timeLastRatedAttributeValue = dynamodb.AttributeValue{}
+	timeLastRatedAttributeValue.SetS(timeLastRated)
+	expressionValuePlaceholders[":timeLastRated"] = &timeLastRatedAttributeValue
+
+	keyMap := make(map[string]*dynamodb.AttributeValue)
+	var key = dynamodb.AttributeValue{}
+	key.SetN(partyID)
+	keyMap["partyID"] = &key
+
+	var updateItemInput = dynamodb.UpdateItemInput{}
+	updateItemInput.SetExpressionAttributeNames(expressionAttributeNames)
+	updateItemInput.SetExpressionAttributeValues(expressionValuePlaceholders)
+	updateItemInput.SetKey(keyMap)
+	updateItemInput.SetTableName("Party")
+	updateExpression := "SET #i.#f.#r=:rating, #i.#f.#t=:timeLastRated"
+	updateItemInput.UpdateExpression = &updateExpression
+	_, err2 := getter.DynamoDB.UpdateItem(&updateItemInput)
+	var dynamodbCall = DynamodbCall{}
+	if err2 != nil {
+		dynamodbCall.Error = "ratePartyHelper function: UpdateItem error (probable cause: this party doesn't exist or you aren't invited to this party). " + err2.Error()
+		dynamodbCall.Succeeded = false
+		queryResult.DynamodbCalls[0] = dynamodbCall
+		return queryResult
+	}
+	dynamodbCall.Succeeded = true
+	queryResult.DynamodbCalls[0] = dynamodbCall
+	queryResult.Succeeded = true
+	return queryResult
+}
+
+func rateBarHelper(barID string, facebookID string, isMale bool, name string, rating string, status string, timeLastRated string) QueryResult {
+	var queryResult = QueryResult{}
+	queryResult.Succeeded = false
+	queryResult.DynamodbCalls = make([]DynamodbCall, 1)
+	type ItemGetter struct {
+		DynamoDB dynamodbiface.DynamoDBAPI
+	}
+	// Setup
+	var getter = new(ItemGetter)
+	var config = &aws.Config{Region: aws.String("us-west-2")}
+	sess, err := session.NewSession(config)
+	if err != nil {
+		queryResult.Error = "rateBarHelper function: session creation error. " + err.Error()
+		return queryResult
+	}
+	var svc = dynamodb.New(sess)
+	getter.DynamoDB = dynamodbiface.DynamoDBAPI(svc)
+	// Finally
+	expressionAttributeNames := make(map[string]*string)
+	attendees := "attendees"
+	expressionAttributeNames["#a"] = &attendees
+	expressionAttributeNames["#f"] = &facebookID
+	expressionValuePlaceholders := make(map[string]*dynamodb.AttributeValue)
+
+	attendeeMap := make(map[string]*dynamodb.AttributeValue)
+	var attendee = dynamodb.AttributeValue{}
+	var isMaleAttribute = dynamodb.AttributeValue{}
+	var nameAttribute = dynamodb.AttributeValue{}
+	var ratingAttribute = dynamodb.AttributeValue{}
+	var statusAttribute = dynamodb.AttributeValue{}
+	var timeLastRatedAttribute = dynamodb.AttributeValue{}
+	isMaleAttribute.SetBOOL(isMale)
+	nameAttribute.SetS(name)
+	ratingAttribute.SetS(rating)
+	statusAttribute.SetS(status)
+	timeLastRatedAttribute.SetS(timeLastRated)
+	attendeeMap["isMale"] = &isMaleAttribute
+	attendeeMap["name"] = &nameAttribute
+	attendeeMap["rating"] = &ratingAttribute
+	attendeeMap["status"] = &statusAttribute
+	attendeeMap["timeLastRated"] = &timeLastRatedAttribute
+	attendee.SetM(attendeeMap)
+	expressionValuePlaceholders[":attendee"] = &attendee
+
+	keyMap := make(map[string]*dynamodb.AttributeValue)
+	var key = dynamodb.AttributeValue{}
+	key.SetN(barID)
+	keyMap["barID"] = &key
+
+	var updateItemInput = dynamodb.UpdateItemInput{}
+	updateItemInput.SetExpressionAttributeNames(expressionAttributeNames)
+	updateItemInput.SetExpressionAttributeValues(expressionValuePlaceholders)
+	updateItemInput.SetKey(keyMap)
+	updateItemInput.SetTableName("Bar")
+	updateExpression := "SET #a.#f=:attendee"
+	updateItemInput.UpdateExpression = &updateExpression
+
+	_, err2 := getter.DynamoDB.UpdateItem(&updateItemInput)
+
+	var dynamodbCall = DynamodbCall{}
+	if err2 != nil {
+		dynamodbCall.Error = "ratePartyHelper function: UpdateItem error. (probable cause: this party may not exist)" + err2.Error()
+		dynamodbCall.Succeeded = false
+		queryResult.DynamodbCalls[0] = dynamodbCall
+		return queryResult
+	}
+	dynamodbCall.Succeeded = true
+	queryResult.DynamodbCalls[0] = dynamodbCall
+	queryResult.Succeeded = true
+	return queryResult
+}
